yolo_demo/requirements.txt
tensorflow>=2.0.0
numpy>=1.18.0
matplotlib>=3.2.0
Pillow>=8.0.0
opencv-python>=4.5.0
Flask>=2.0.0